move default uam permission toml to config/
