+++ pkg/common/uam_permissions_config.go
@@ -1,4 +1,4 @@
-package config
+package common
 
 import (
 	"fmt"
